# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Contract Management System API
  description: |
    Complete API documentation for the Contract Management System.
    
    ## Authentication
    All API requests require authentication using JWT tokens.
    Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Anonymous requests: 20/hour
    - Authenticated requests: 1000/hour
    - Premium users: 5000/hour
    
    ## WebSocket Support
    Real-time updates are available via WebSocket connections at:
    - Production: wss://api.contractmanagement.com/ws
    - Development: ws://localhost:8000/ws
    
  version: 1.0.0
  contact:
    name: API Support
    email: api@contractmanagement.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.contractmanagement.com/v1
    description: Production server
  - url: https://staging-api.contractmanagement.com/v1
    description: Staging server
  - url: http://localhost:8000/api/v1
    description: Development server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contractId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        content:
          type: string
        parentId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /contracts/{id}/comments:
    get:
      tags:
        - Contracts
      summary: Get contract comments
      operationId: getContractComments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Comments retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
                      
    post:
      tags:
        - Contracts
      summary: Add comment to contract
      operationId: addContractComment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                parentId:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        201:
          description: Comment added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
...